separate out simplification functionality
