+++ src/nlsat/CMakeLists.txt
@@ -4,6 +4,7 @@ z3_add_component(nlsat
     nlsat_evaluator.cpp
     nlsat_explain.cpp
     nlsat_interval_set.cpp
+    nlsat_simplify.cpp
     nlsat_solver.cpp
     nlsat_types.cpp
   COMPONENT_DEPENDENCIES
